Code Cleanup
- renamed folder `Assets/SEE/Tools/Videochat/` to `Livekit`

- LivekitVideo.cs
  - edit dokumentation
  - added Debug.LogError if the playerHead cannot be found and then disable the component.
  - added namespace
  - remove unnecessary query in `Update()`
  - `var` has been removed

- LivekitVideoManager.cs
  - `var` has been removed
  - added namespace
  - added #Region
  - added capital letters for public fields
  - removed `_` for identifiers
  - simplified new operator
  - fixed spelling and Grammar
  - fixed indentation
